se reorganizo backend y se creo el frontend

+++ backend/src/controller/controller_user.py
@@ -5,13 +5,14 @@ from datetime import date
 from typing import List
 from src.controller.controoler_auth import get_password_hash
 from src.models.models_user import User
+import aiofiles as aofiles #se usa para manejar archivos de forma asincrona
 
 #directorios para guardar archivos e img
-Upload_dir="public/docs"
+Upload_dir="public/docs/user"
 FILE_MAXIMUN=5
 MAX_FILE_SIZE = 20 * 1024 * 1024  
 TYPE_FILE=["application/pdf","application/png","image/jpeg","image/jpg"]
-UPLOAD_DIR="public/img"
+UPLOAD_DIR="public/img/user"
 
 def get_user():
     try:
@@ -79,9 +80,7 @@ def create_user(model: User, imagen: UploadFile, archivo: List[UploadFile]):
         else:
             print("⚠️ No se recibió ninguna imagen")
 
-        # -----------------------------------
         # GUARDAR ARCHIVOS
-        # -----------------------------------
         # Normalizar archivos (si solo viene uno)
         if not archivo:
             archivo = []
@@ -96,7 +95,6 @@ def create_user(model: User, imagen: UploadFile, archivo: List[UploadFile]):
             os.makedirs(Upload_dir, exist_ok=True)
 
             for f in archivo:
-
                 if getattr(f, "content_type", None) not in TYPE_FILE:
                     raise HTTPException(
                         status_code=415,
@@ -129,10 +127,11 @@ def create_user(model: User, imagen: UploadFile, archivo: List[UploadFile]):
         # -----------------------------------
         clave= model.clave[:72]
         password_finish=get_password_hash(clave)
+        estado_default="Activo"
         cursor.execute(
-            "INSERT INTO usuarios (identificacion, nombre, correo, edad, descripcion, rol, clave, imagen) "
-            "VALUES (%s, %s, %s, %s, %s, %s, %s, %s)",
-            (model.identificacion, model.nombre, model.correo, model.edad, model.descripcion, model.rol, password_finish, imagen_filename)
+            "INSERT INTO usuarios (identificacion, nombre, correo, edad, descripcion, rol, clave, imagen, estado) "
+            "VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)",
+            (model.identificacion, model.nombre, model.correo, model.edad, model.descripcion, model.rol, password_finish, imagen_filename, estado_default)
         )
 
         if name_original:
@@ -153,88 +152,76 @@ def create_user(model: User, imagen: UploadFile, archivo: List[UploadFile]):
         raise HTTPException(status_code=500, detail=str(ex))
 
     
-#falta  quee encripte la clave al atcualizar
-def update_user(user_id, user, imagen, archivos):
+
+async def update_user(user_id, user:User, imagen, archivos):
     try:
         connection = get_connection()
-        # Usa dictionary=True aquí si quieres acceder por nombre, si no, usa índices como ya hacías
-        cursor = connection.cursor(dictionary=False) 
-        cursor.execute("SELECT * FROM usuarios WHERE identificacion=%s", (user_id,))
-        existing_user = cursor.fetchone()
-        
+        cursor = connection.cursor(dictionary=True)
+        cursor.execute("select  usuarios.identificacion, usuarios.nombre as nombre, usuarios.correo, usuarios.clave, usuarios.edad, usuarios.descripcion, usuarios.rol, usuarios.imagen, usuarios.estado, usuarios.estado, archivos.nombre as nombre_archivo, archivos.id as id_archivo from usuarios join archivos on usuarios.identificacion = archivos.id_usuario where usuarios.identificacion=%s", (user_id,)) 
+        existing_user= cursor.fetchone()
+
         if not existing_user:
-            raise HTTPException(status_code=404, detail="Usuario no encontrado")
+            raise HTTPException(status_code=404, detail="Usuario no encontrado para actualizar.")
+           
+        nombre= user.nombre if user.nombre is not None else existing_user["nombre"]
+        correo= user.correo if user.correo is not None else existing_user["correo"]
+        clave= user.clave if user.clave is not None else existing_user["clave"]
+        edad= user.edad if user.edad is not None else existing_user["edad"]
+        descripcion= user.descripcion if user.descripcion is not None else existing_user["descripcion"]
+        rol= user.rol if user.rol is not None else existing_user["rol"]
+        estado= user.estado if user.estado is not None else existing_user["estado"]
 
-        # Mapear valores actuales / nuevos, manejando strings vacíos si es necesario
-        nombre_final = user.nombre if user.nombre is not None and user.nombre.strip() != "" else existing_user[1]
-        correo_final = user.correo if user.correo is not None and user.correo.strip() != "" else existing_user[2]
-        
-        # --- Lógica de Encriptación ---
-        if clave not in [None, "", " "]:
-            clave_final = get_password_hash(user.clave[:72])
-            print("🔒 Nueva clave encriptada:", clave_final)
-        else:
-            print("ℹ️ Clave no enviada, se mantiene la existente")
+        if clave != existing_user["clave"]:
+            clave= clave[:72]
+            clave=get_password_hash(clave)
 
-
-        edad_final = user.edad if user.edad is not None else existing_user[4]
-        descripcion_final = user.descripcion if user.descripcion is not None else existing_user[5]
-        rol_final = user.rol if user.rol is not None else existing_user[6]
-        imagen_final = existing_user[7]
-
-        # Si hay imagen nueva
+        imagen_filename = None
         if imagen:
             os.makedirs(UPLOAD_DIR, exist_ok=True)
+            if existing_user["imagen"]:
+                nombre_img= existing_user["imagen"]
+                ruta_img=os.path.join(UPLOAD_DIR, nombre_img)
+                os.remove(ruta_img)
+
             imagen_filename = imagen.filename
             imagen_path = os.path.join(UPLOAD_DIR, imagen_filename)
-            with open(imagen_path, "wb") as buffer:
+            async with aofiles.open(imagen_path, "wb") as buffer:
                 buffer.write(imagen.file.read())
             print("✅ Imagen guardada:", imagen_filename)
-            imagen_final = imagen_filename
+        else:
+            print("⚠️ No se recibió ninguna imagen")
 
-        # Ejecutar UPDATE
-        cursor.execute(
-            """
-            UPDATE usuarios
-            SET nombre=%s, correo=%s, clave=%s, edad=%s, descripcion=%s, rol=%s, imagen=%s
-            WHERE identificacion=%s
-            """,
-            (nombre_final, correo_final, clave_final, edad_final, descripcion_final, rol_final, imagen_final, user_id)
-        )
-
-        # Guardar archivos si vienen
-        if archivos:
+        if archivos and len(archivos) > 0:
             os.makedirs(Upload_dir, exist_ok=True)
-            for f in archivos:
-                if getattr(f, "content_type", None) not in TYPE_FILE:
-                    raise HTTPException(status_code=415, detail=f"Tipo no permitido: {f.content_type}")
 
-                name_original = f.filename
-                dest_path = os.path.join(Upload_dir, name_original)
+            cursor.execute("select id, nombre FROM archivos WHERE id_usuario=%s", (user_id,))
+            data_file= cursor.fetchall()
+            #se encarga de eliminar los archivos viejos
+            for file_db in data_file:
+                ruta= os.path.join(Upload_dir, file_db["nombre"])
+                if os.path.exists(ruta):
+                    os.remove(ruta)
+                cursor.execute("DELETE FROM archivos WHERE id=%s",(file_db["id"],))
+            #recibe cada archivo y lo guarda en la carpeta
+            for archivo in archivos:
+                nombre_file= archivo.filename   
+                dest_path= os.path.join(Upload_dir, nombre_file)
+                async with aofiles.open(dest_path, "wb") as buffer:
+                    buffer.write(archivo.file.read())
+                tipo_archivo= archivo.content_type  
+                fecha_carga= date.today()
+                
+                cursor.execute("INSERT INTO archivos (fecha_carga, tipo_archivo, nombre, id_usuario) VALUES (%s, %s, %s, %s)",
+                (fecha_carga, tipo_archivo, nombre_file, user_id))
 
-                total = 0
-                with open(dest_path, "wb") as out:
-                    while chunk := f.file.read(1024 * 64):
-                        out.write(chunk)
-                        total += len(chunk)
-                        if total > MAX_FILE_SIZE:
-                            out.close()
-                            os.remove(dest_path)
-                            raise HTTPException(status_code=413, detail=f"Archivo demasiado grande: {f.filename}")
 
-                tipo_archivo = f.content_type
-                fecha_carga = date.today()
-
-                cursor.execute(
-                    "INSERT INTO archivos (fecha_carga, tipo_archivo, nombre, id_usuario, id_finca) VALUES (%s, %s, %s, %s)",
-                    (fecha_carga, tipo_archivo, name_original, user_id)
-                )
-
-                print("✅ Archivo guardado:", name_original)
-
-        connection.commit()
-        return {"status": 200, "mensaje": "Usuario actualizado exitosamente"}
-
+            cursor.execute("update usuarios set nombre=%s, correo=%s, clave=%s, edad=%s, descripcion=%s, rol=%s, imagen=%s, estado=%s where identificacion=%s",
+                           (nombre, correo, clave, edad, descripcion, rol, imagen_filename, estado,user_id))
+            connection.commit()
+            if cursor.rowcount >= 0: 
+                return {"status": 200, "mensaje": "Usuario actualizado exitosamente"}
+            else:
+                raise HTTPException(status_code=400, detail="No se pudo actualizar el usuario.")
     except Exception as ex:
         print("Error al actualizar el usuario:", ex)
         raise HTTPException(status_code=500, detail=str(ex))
@@ -242,18 +229,28 @@ def update_user(user_id, user, imagen, archivos):
 
 def  delete_user(user_id: int):
     try:
+        
         #conecta a la base de datos
         connect= get_connection()
         #crear un cursor  para ejecutar consultas
         cursor= connect.cursor(dictionary=True)
-        #ejecuta la consulta
-        cursor.execute("DELETE FROM usuarios WHERE identificacion=%s", (user_id,))
-        #obtener el resultado y no se puede quedar sin esto si no no borra
-        connect.commit()
-        if cursor.rowcount == 0:
-            raise HTTPException(status_code=404, detail="Usuario no encontrado")
-        return {"status": 200, "mensaje": "Usuario eliminado exitosamente"}
-    #este es como el catch en node js y atrapa el error 
+        cursor.execute("select id, nombre FROM archivos WHERE id_usuario=%s", (user_id,))
+        data_file= cursor.fetchall()
+        if len(data_file)==0: 
+            raise HTTPException(status_code=404, detail="Usuario no encontrado o archivo para eliminar verifica tu identificacion.")
+
+        for file_db in data_file:
+            ruta= os.path.join(Upload_dir, file_db["nombre"])
+            if os.path.exists(ruta):
+                os.remove(ruta)
+            cursor.execute("DELETE FROM archivos WHERE id=%s",(file_db["id"],))
+            if cursor.rowcount >= 0:
+                cursor.execute("DELETE FROM usuarios WHERE identificacion=%s", (user_id,))
+                connect.commit()
+                if cursor.rowcount == 0:
+                    raise HTTPException(status_code=404, detail="Usuario no encontrado")
+                return {"status": 200, "mensaje": "Usuario eliminado exitosamente"}
+             
     except Exception as ex:
         print("Error al obtener el usuario:", ex)
         raise HTTPException(status_code=500, detail=str(ex))
